requirements.txt
httpx>=0.27.0
pandas>=2.0.0
xlrd>=2.0.1
openpyxl>=3.1.0
PyMuPDF>=1.24.0
google-cloud-vision>=3.7.0
openai>=1.40.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
